@@ -14,64 +14,59 @@
     }
   },
   "scripts": {
-    "clean": "rm dist/bundle.js",
-    "build": "node-sass ./stylesheets -o ./dist/stylesheets --output-style compressed && webpack --mode=production",
-    "build-css": "node-sass ./stylesheets -o ./dist/stylesheets --output-style compressed",
+    "build": "node-sass ./stylesheets -o ./dist/stylesheets --output-style compressed && webpack --config webpack.config.js",
     "build-dev": "webpack -d --mode development -w",
     "build-prod": "webpack -p --mode production -w",
-    "watch-css": "node-sass --watch ./stylesheets -o ./dist/stylesheets --output-style compressed",
     "test": "jest",
     "test-cov": "jest --collect-coverage",
-    "start": "node ./server/index.js",
-    "start-dev": "nodemon ./server/index.js"
+    "start": "node ./index.js",
+    "start-dev": "nodemon ./index.js"
   },
   "dependencies": {
-    "axios": "^0.20.0",
+    "axios": "^0.21.0",
     "compression": "^1.7.4",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "jquery": "^3.5.1",
     "lodash": "^4.17.20",
-    "moment": "^2.28.0",
+    "moment": "^2.29.1",
     "path": "^0.12.7",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-redux": "^7.2.1",
-    "react-router": "^5.2.0",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0"
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "react-redux": "^7.2.2",
+    "react-router-dom": "^5.2.0",
+    "redux": "^4.0.5"
   },
   "devDependencies": {
-    "@babel/core": "^7.11.6",
-    "@babel/preset-env": "^7.11.5",
-    "@babel/preset-react": "^7.10.4",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.0.4",
-    "babel-jest": "^26.3.0",
-    "babel-loader": "^8.1.0",
-    "brotli-webpack-plugin": "^1.1.0",
-    "compression-webpack-plugin": "^6.0.2",
-    "css-loader": "^4.3.0",
-    "eslint": "^7.9.0",
-    "eslint-config-prettier": "^6.11.0",
+    "@babel/core": "^7.12.9",
+    "@babel/preset-env": "^7.12.7",
+    "@babel/preset-react": "^7.12.7",
+    "@testing-library/jest-dom": "^5.11.6",
+    "@testing-library/react": "^11.2.2",
+    "babel-jest": "^26.6.3",
+    "babel-loader": "^8.2.2",
+    "compression-webpack-plugin": "^6.1.1",
+    "css-loader": "^5.0.1",
+    "eslint": "^7.14.0",
+    "eslint-config-prettier": "^6.15.0",
     "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.20.6",
+    "eslint-plugin-react": "^7.21.5",
     "html-critical-webpack-plugin": "^2.1.0",
     "html-webpack-plugin": "^4.5.0",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^26.4.2",
+    "jest": "^26.6.3",
     "jest-dom": "^4.0.0",
     "lodash-webpack-plugin": "^0.11.5",
-    "mini-css-extract-plugin": "^0.12.0",
-    "node-sass": "^4.14.1",
-    "nodemon": "^2.0.4",
-    "prettier": "^2.1.2",
+    "mini-css-extract-plugin": "^1.3.1",
+    "node-sass": "^5.0.0",
+    "nodemon": "^2.0.6",
+    "prettier": "^2.2.1",
     "react-test-render": "^1.1.2",
-    "sass-loader": "^10.0.2",
-    "style-loader": "^1.2.1",
-    "terser-webpack-plugin": "^4.2.3",
-    "webpack": "^4.44.2",
-    "webpack-cli": "^3.3.12"
+    "sass-loader": "^10.1.0",
+    "style-loader": "^2.0.0",
+    "terser-webpack-plugin": "^5.0.3",
+    "webpack": "^5.9.0",
+    "webpack-cli": "^4.2.0"
   },
   "engines": {
     "node": "12.18.3"
