@@ -1,24 +1,10 @@
-require('dotenv').config();
 const axios = require('axios');
-const express = require('express');
-const bodyParser = require('body-parser');
+const express = require('express')
+const ROUTER = express.Router();
 const path = require('path');
-const compression = require('compression');
-const APP = express();
-const PORT = process.env.PORT || 3005;
 const APIURL = process.env.API_URL;
 
-APP.use(bodyParser.json());
-APP.use(bodyParser.urlencoded({ extended: true }));
-APP.use(compression())
-APP.use(express.static(path.join(__dirname, '../dist')));
-
-APP.get('/', (req, res) => {
-  res.set('Content-Encoding', 'gzip')
-  res.status(200).send('./index.html')
-});
-
-APP.get('/catwalk/:id', (req, res) => {
+ROUTER.get('/:id', (req, res) => {
   var id = req.params.id
   const params = [
     axios.get(`${APIURL}/products/${id}`),
@@ -46,7 +32,7 @@ APP.get('/catwalk/:id', (req, res) => {
     })
 });
 
-APP.post('/catwalk/interactions', (req, res) => {
+ROUTER.post('/interactions', (req, res) => {
   return axios.post(`${APIURL}/interactions`, req.body)
     .then((response) => {
       res.send('interaction posted');
@@ -59,7 +45,7 @@ APP.post('/catwalk/interactions', (req, res) => {
 
 //Q&A ROUTES//
 
-APP.get('/catwalk/qa/questions/:id', (req, res) => {
+ROUTER.get('/qa/questions/:id', (req, res) => {
   var id = req.params.id
   axios.get(`${APIURL}/qa/questions/?product_id=${id}&count=50`)
     .then((response) => {
@@ -71,7 +57,7 @@ APP.get('/catwalk/qa/questions/:id', (req, res) => {
     })
 });
 
-APP.post('/catwalk/qa/questions', (req, res) => {
+ROUTER.post('/qa/questions', (req, res) => {
   axios.post(`${APIURL}/qa/questions`, req.body)
     .then((response) => {
       res.send('question posted');
@@ -82,7 +68,7 @@ APP.post('/catwalk/qa/questions', (req, res) => {
     })
 });
 
-APP.put('/catwalk/qa/questions/helpful/:id', (req, res) => {
+ROUTER.put('/qa/questions/helpful/:id', (req, res) => {
   var questionId = req.params.id;
   return axios.put(`${APIURL}/qa/questions/${questionId}/helpful`)
     .then((response) => {
@@ -94,7 +80,7 @@ APP.put('/catwalk/qa/questions/helpful/:id', (req, res) => {
     })
 });
 
-APP.put('/catwalk/qa/questions/report/:id', (req, res) => {
+ROUTER.put('/qa/questions/report/:id', (req, res) => {
   var questionId = req.params.id;
   return axios.put(`${APIURL}/qa/questions/${questionId}/report`)
     .then((response) => {
@@ -106,7 +92,7 @@ APP.put('/catwalk/qa/questions/report/:id', (req, res) => {
     })
 });
 
-APP.post('/catwalk/qa/answers/:id', (req, res) => {
+ROUTER.post('/qa/answers/:id', (req, res) => {
   var questionId = req.params.id
   return axios.post(`${APIURL}/qa/questions/${questionId}/answers`, req.body)
     .then((response) => {
@@ -118,7 +104,7 @@ APP.post('/catwalk/qa/answers/:id', (req, res) => {
     })
 });
 
-APP.put('/catwalk/qa/answers/helpful/:id', (req, res) => {
+ROUTER.put('/qa/answers/helpful/:id', (req, res) => {
   var answerId = req.params.id;
   return axios.put(`${APIURL}/qa/answers/${answerId}/helpful`)
     .then((response) => {
@@ -130,7 +116,7 @@ APP.put('/catwalk/qa/answers/helpful/:id', (req, res) => {
     })
 });
 
-APP.put('/catwalk/qa/answers/report/:id', (req, res) => {
+ROUTER.put('/qa/answers/report/:id', (req, res) => {
   var answerId = req.params.id;
   return axios.put(`${APIURL}/qa/answers/${answerId}/report`)
     .then((response) => {
@@ -142,6 +128,4 @@ APP.put('/catwalk/qa/answers/report/:id', (req, res) => {
     })
 });
 
-APP.listen(PORT, () => {
-  console.log(`listening on port ${PORT}`);
-});
+module.exports = ROUTER;
