@@ -1,46 +1,19 @@
 import axios from 'axios';
-import { createStore, applyMiddleware } from 'redux';
-import thunk from 'redux-thunk';
+import { createStore } from 'redux';
 import rootReducer from './../reducers/main.js';
 
-var store = {};
-
-function initializeStore(id, callback) {
-  return axios
-    .get(`/catwalk/${id}`)
-    .then((response) => {
-      const data = response.data;
-      const styleData = data.primaryProductStyles.results
-      styleData.forEach((style) => {
-        style.photos.forEach((photo) => {
-          if (photo.thumbnail_url === null) {
-            photo.thumbnail_url = './attributes/no-img.png';
-          }
-          if (photo.url === null) {
-            photo.url = './attributes/no-img.png';
-          }
-        })
-      })
-      var defaultState = {
-        primaryProduct: data.primaryProduct,
-        currentStyles: styleData,
-        currentStyle: styleData[0],
-        currentImages: styleData[0].photos,
-        currentImage: styleData[0].photos[0],
-        productQuestions: data.primaryProductQuestions.results,
-        sortingMethod: 'relevance',
-        reviews: data.primaryProductReviews.results,
-        primaryProductMetadata: data.primaryProductReviewsNumbers,
-      };
-
-      store = createStore(rootReducer, defaultState, applyMiddleware(thunk));
-    })
-    .then(() => {
-      callback();
-    })
-    .catch((err) => {
-      console.log('error creating the store', err);
-    });
+const defaultState = {
+  primaryProduct: null,
+  currentStyles: null,
+  currentStyle:  null,
+  currentImages: null,
+  currentImage: null,
+  relatedProducts: null,
+  productQuestions: null,
+  reviews: null,
+  primaryProductMetadata: null
 }
 
-export { store, initializeStore };
+const store = createStore(rootReducer, defaultState);
+
+export default store;
